refactor: 메서드 및 패키지 명 변경
